Game Over+New Start
made it so that if you go into the green trail it’s a game over, the
green is sort of an enemy, but I’m thinking of putting in another
moving enemy. Level one is pretty much done, get to the yellow to get
to level 2 without walking in the green trail made. if you go in the
green, there’s a noise and you restart.